+++ test/unit/dropin-model.js
@@ -142,6 +142,44 @@ describe('DropinModel', () => {
   });
 
   describe('initialize', () => {
+    test('emits asyncDependenciesReady event when no dependencies are set to initializing', (done) => {
+      jest.useFakeTimers();
+
+      const model = new DropinModel(testContext.modelOptions);
+
+      jest.spyOn(model, '_emit');
+
+      model.on('asyncDependenciesReady', () => {
+        jest.useRealTimers();
+
+        done();
+      });
+
+      return model.initialize().then(() => {
+        expect(model._emit).not.toBeCalledWith('asyncDependenciesReady');
+
+        model.asyncDependencyReady('paypal');
+
+        expect(model._emit).not.toBeCalledWith('asyncDependenciesReady');
+
+        jest.advanceTimersByTime(1000);
+
+        expect(model._emit).not.toBeCalledWith('asyncDependenciesReady');
+
+        model.asyncDependencyReady('venmo');
+
+        expect(model._emit).not.toBeCalledWith('asyncDependenciesReady');
+
+        jest.advanceTimersByTime(1000);
+
+        model.asyncDependencyReady('card');
+
+        expect(model._emit).not.toBeCalledWith('asyncDependenciesReady');
+
+        jest.advanceTimersByTime(1000);
+      });
+    });
+
     test('creates a vault manager', () => {
       const model = new DropinModel(testContext.modelOptions);
 
@@ -187,6 +225,7 @@ describe('DropinModel', () => {
         expect(model._paymentMethods).toEqual([
           { type: 'CreditCard', details: { lastTwo: '11' }, vaulted: true }
         ]);
+        expect(model.dependencyStates.paypal).toBe('not-enabled');
       });
     });
 
@@ -579,6 +618,38 @@ describe('DropinModel', () => {
     });
   });
 
+  describe('hasAtLeastOneAvailablePaymentOption', () => {
+    test('returns false when no payment methods have a setup status of done', async () => {
+      // all dependencies will have a setup status of initializing
+      const model = new DropinModel(testContext.modelOptions);
+
+      await model.initialize();
+
+      expect(model.hasAtLeastOneAvailablePaymentOption()).toBe(false);
+    });
+
+    test('returns true when at least one payment method has a setup status of done', async () => {
+      const model = new DropinModel(testContext.modelOptions);
+
+      await model.initialize();
+
+      model.asyncDependencyReady('card');
+
+      expect(model.hasAtLeastOneAvailablePaymentOption()).toBe(true);
+    });
+
+    test('returns false when no payment methods have a setup status of done, even if non-payment methods do', async () => {
+      const model = new DropinModel(testContext.modelOptions);
+
+      await model.initialize();
+
+      model.asyncDependencyReady('dataCollector');
+      model.asyncDependencyReady('threeDSecure');
+
+      expect(model.hasAtLeastOneAvailablePaymentOption()).toBe(false);
+    });
+  });
+
   describe('reportAppSwitchPayload', () => {
     test('saves app switch payload to instance', () => {
       const model = new DropinModel(testContext.modelOptions);
@@ -602,22 +673,20 @@ describe('DropinModel', () => {
     });
   });
 
-  describe('asyncDependencyStarting', () => {
-    beforeEach(() => {
-      testContext.context = {
-        dependenciesInitializing: 0
-      };
-    });
-
-    test('increments dependenciesInitializing by one', () => {
-      DropinModel.prototype.asyncDependencyStarting.call(testContext.context);
-      expect(testContext.context.dependenciesInitializing).toBe(1);
-    });
-  });
-
   describe('asyncDependencyFailed', () => {
     beforeEach(() => {
       testContext.model = new DropinModel(testContext.modelOptions);
+    });
+
+    test('marks depenedency as failed', () => {
+      const err = new Error('a bad error');
+
+      expect(testContext.model.dependencyStates.venmo).toBe('initializing');
+      testContext.model.asyncDependencyFailed({
+        view: 'venmo',
+        error: err
+      });
+      expect(testContext.model.dependencyStates.venmo).toBe('failed');
     });
 
     test('adds an error to the failedDependencies object', () => {
@@ -656,10 +725,13 @@ describe('DropinModel', () => {
           done();
         });
 
-        model.asyncDependencyStarting();
-        model.asyncDependencyFailed({
-          view: 'id',
-          error: new Error('fake error')
+        model.initialize().then(() => {
+          model.asyncDependencyReady('paypal');
+          model.asyncDependencyFailed({
+            view: 'venmo',
+            error: new Error('fake error')
+          });
+          model.asyncDependencyReady('card');
         });
       }
     );
@@ -670,14 +742,27 @@ describe('DropinModel', () => {
       testContext.context = { callback: jest.fn() };
     });
 
-    test('decrements dependenciesInitializing by one', () => {
+    test('marks initializing dependency as done', () => {
       const model = new DropinModel(testContext.modelOptions);
 
-      model.dependenciesInitializing = 2;
+      // the modelOptions have a config for paypal and venmo
+      expect(model.dependencyStates.paypal).toBe('initializing');
+      expect(model.dependencyStates.venmo).toBe('initializing');
 
-      model.asyncDependencyReady();
+      model.asyncDependencyReady('paypal');
+      expect(model.dependencyStates.paypal).toBe('done');
+      expect(model.dependencyStates.venmo).toBe('initializing');
+      expect(model.dependencyStates.card).toBe('initializing');
 
-      expect(model.dependenciesInitializing).toBe(1);
+      model.asyncDependencyReady('venmo');
+      expect(model.dependencyStates.paypal).toBe('done');
+      expect(model.dependencyStates.venmo).toBe('done');
+      expect(model.dependencyStates.card).toBe('initializing');
+
+      model.asyncDependencyReady('card');
+      expect(model.dependencyStates.paypal).toBe('done');
+      expect(model.dependencyStates.venmo).toBe('done');
+      expect(model.dependencyStates.card).toBe('done');
     });
 
     test(
@@ -688,30 +773,34 @@ describe('DropinModel', () => {
         jest.spyOn(DropinModel.prototype, 'asyncDependencyReady');
 
         model.on('asyncDependenciesReady', () => {
-          expect(DropinModel.prototype.asyncDependencyReady).toBeCalledTimes(1);
+          expect(DropinModel.prototype.asyncDependencyReady).toBeCalledTimes(3);
           done();
         });
 
-        model.asyncDependencyStarting();
-        model.asyncDependencyReady();
+        model.initialize().then(() => {
+          model.asyncDependencyReady('paypal');
+          model.asyncDependencyReady('venmo');
+          model.asyncDependencyReady('card');
+        });
       }
     );
 
-    test('emits asyncDependenciesReady event with prior errors', () => {
+    test('emits asyncDependenciesReady event with prior errors', (done) => {
       const model = new DropinModel(testContext.modelOptions);
       const err = new Error('an earlier dependency failed');
 
-      jest.spyOn(model, '_emit');
-
-      model.asyncDependencyStarting();
-      model.asyncDependencyStarting();
-      model.asyncDependencyFailed({
-        view: 'id',
-        error: err
+      model.on('asyncDependenciesReady', () => {
+        done();
       });
-      model.asyncDependencyReady();
 
-      expect(model._emit).toBeCalledWith('asyncDependenciesReady');
+      return model.initialize().then(() => {
+        model.asyncDependencyReady('card');
+        model.asyncDependencyFailed({
+          view: 'venmo',
+          error: err
+        });
+        model.asyncDependencyReady('paypal');
+      });
     });
   });